backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
nest_asyncio
